:recycle: migrate tests about VStack children position
